chore: split assembler.rs budget code into budget_assembly.rs
Extract ~152 lines of budget-related code (budget_for, active_turn_budget,
helpers, BudgetBackend tests) from context/assembler.rs into a new module
context/budget_assembly.rs.  assembler.rs drops from 954 to 802 lines.
